requirements.txt
bleak>=0.21.0
websockets>=12.0
pycryptodome>=3.19.0
